pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-project-template"
version = "0.1.0"
description = "A production-ready Python project template"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "pytest>=7.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.3.0",
    "isort>=5.10.1",
    "flake8>=4.0.1",
    "mypy>=0.950",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3 